+++ ios/RNImgly.h
@@ -1,0 +1,1 @@
+#define RN_IMGLY RNPESDK // Product prefix